+++ src/webparts/controlsTest/components/ControlsTest.tsx
@@ -17,6 +17,7 @@ import { IFrameDialog } from '../../../IFrameDialog';
 import { Environment, EnvironmentType } from '@microsoft/sp-core-library';
 import { SecurityTrimmedControl, PermissionLevel } from '../../../SecurityTrimmedControl';
 import { SPPermission } from '@microsoft/sp-page-context';
+import { PeoplePicker } from '../../../PeoplePicker';
 
 /**
  * Component that can be used to test out the React controls from this project
@@ -105,6 +106,12 @@ export default class ControlsTest extends React.Component<IControlsTestProps, IC
         items: items
       });
     }
+  }
+  /** Method that retrieves the selected items from People  Picker
+   * @param items
+   */
+  private _getPeoplePickerItems(items: any[]) {
+    console.log('Items:', items);
   }
 
   /**
@@ -293,6 +300,15 @@ export default class ControlsTest extends React.Component<IControlsTestProps, IC
           selection={this._getSelection} />
 
           <p><a href="javascript:;" onClick={this.deleteItem}>Deletes second item</a></p>
+
+          <PeoplePicker
+            context={this.props.context}
+            titleText="People Picker"
+            // personSelectionLimit={3}
+            // groupName={"Team Site Owners"}
+            showtooltip={true}
+            isRequired={true}
+            selectedItems={this._getPeoplePickerItems} />
       </div>
     );
   }